% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_cover_ml.R
\name{canopy_cover_ml}
\alias{canopy_cover_ml}
\title{canopy_cover_ml
Wrapper function for canopy cover calculation using imageseg package, a machine learning method of canopy cover analysis}
\usage{
canopy_cover_ml(imageDir, outDir_resized, outDir_binary, canopy_model)
}
\arguments{
\item{imageDir}{directory to images to analyze}

\item{outDir_resized}{directory to resized images}

\item{outDir_binary}{directory to binarized images}

\item{canopy_model}{filepath to canopy model downloaded from imageseg (work on integrating this into the package so download isn't necessary)}
}
\description{
canopy_cover_ml
Wrapper function for canopy cover calculation using imageseg package, a machine learning method of canopy cover analysis
}
\details{
Follow instructions in imageseg documentation to ensure keras is properly installed. Canopy model should be downloaded from imageseg and saved to be provided to the function
}
