\name{ConvertSparseList2Array}
\alias{ConvertSparseList2Array}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Converts a Sparse List to an Array
}
\description{
see exampke in Rmarkdown
}
\usage{
ConvertSparseList2Array(OneObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{OneObject}{
a sparse list element with matrix defined by package Matrix
}
}
\value{
[1:n,1:d,1:3] array of numerical entries
}
\author{
Michael Thrun
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
\code{\link{ConvertArray2SparseList}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
