\name{ImageProcessing-package}
\alias{ImageProcessing-package}
\alias{ImageProcessing}
\docType{package}
\title{
\packageTitle{ImageProcessing}
}
\description{
\packageDescription{ImageProcessing}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ImageProcessing}
\packageIndices{ImageProcessing}
}
\author{
\packageAuthor{ImageProcessing}

Maintainer: \packageMaintainer{ImageProcessing}
}
\references{
[Stoll et al., 2015]  Stoll, J., Thrun, M. C., Nuthmann, A., & Einhäuser, W.: Overt attention in natural scenes: objects dominate features, Vision research, Vol. 107, pp. 36-48, doi 10.1016/j.visres.2014.11.006, 2015.


}

\keyword{ImageProcessing}
\seealso{
..
}
\examples{
#todo
}
