% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_area.R
\name{leaf_area}
\alias{leaf_area}
\title{Leaf area
Calculate leaf area using pliman package}
\usage{
leaf_area(imgpath, bg_file, area, dir_processed, outdir = NA, ...)
}
\arguments{
\item{imgpath}{vector of file paths to images}

\item{bg_file}{file path to an image of the background behind leaves (see details in pliman package)}

\item{area}{area of reference square}

\item{dir_processed}{directory to save processed images}

\item{outdir}{if specified, filepath to save output dataframe}

\item{...}{arguments supplied to analyze_objects() from pliman package}
}
\description{
Leaf area
Calculate leaf area using pliman package
}
\details{
Reference area must be a square. It is recommended that you follow instructions in pliman package to identify the best index for your analysis.
The default used here is "NB". From testing, this appears to work well for a green leaf on a royal blue background with a white reference square.
For more instructions on calculating leaf area, see pliman reference at https://tiagoolivoto.github.io/pliman/
}
